methods are updated

+++ src/main/java/com/app/candidate/Repository/CandidateRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CandidateRepository extends JpaRepository<Candidate,Long> {
+public interface CandidateRepository extends JpaRepository<Candidate,Integer> {
 }
